refactor: enforce package boundaries in sva studio react
