+++ apps/sva-studio-react/src/lib/mainserver-events-api.server.ts
@@ -1,0 +1,3 @@
+import { dispatchSvaMainserverEventsRequest } from '@sva/sva-mainserver/server';
+
+export const dispatchMainserverEventsRequest = dispatchSvaMainserverEventsRequest;
