+++ apps/sva-studio-react/src/lib/mainserver-poi-api.server.ts
@@ -1,0 +1,3 @@
+import { dispatchSvaMainserverPoiRequest } from '@sva/sva-mainserver/server';
+
+export const dispatchMainserverPoiRequest = dispatchSvaMainserverPoiRequest;
