+++ apps/sva-studio-react/src/server.test.ts
@@ -4,7 +4,8 @@ const createStartHandlerMock = vi.fn();
 const createSdkLoggerMock = vi.fn();
 const dispatchAuthRouteRequestMock = vi.fn();
 const dispatchMainserverNewsRequestMock = vi.fn();
-const dispatchMainserverEventsPoiRequestMock = vi.fn();
+const dispatchMainserverEventsRequestMock = vi.fn();
+const dispatchMainserverPoiRequestMock = vi.fn();
 const getWorkspaceContextMock = vi.fn();
 const withRequestContextMock = vi.fn();
 const createServerFunctionRequestDiagnosticsMock = vi.fn();
@@ -34,8 +35,12 @@ vi.mock('./lib/mainserver-news-api.server', () => ({
   dispatchMainserverNewsRequest: dispatchMainserverNewsRequestMock,
 }));
 
-vi.mock('./lib/mainserver-events-poi-api.server', () => ({
-  dispatchMainserverEventsPoiRequest: dispatchMainserverEventsPoiRequestMock,
+vi.mock('./lib/mainserver-events-api.server', () => ({
+  dispatchMainserverEventsRequest: dispatchMainserverEventsRequestMock,
+}));
+
+vi.mock('./lib/mainserver-poi-api.server', () => ({
+  dispatchMainserverPoiRequest: dispatchMainserverPoiRequestMock,
 }));
 
 vi.mock('./lib/server-function-request-diagnostics.server', () => ({
@@ -53,7 +58,8 @@ describe('server transport', () => {
     createSdkLoggerMock.mockReset();
     dispatchAuthRouteRequestMock.mockReset();
     dispatchMainserverNewsRequestMock.mockReset();
-    dispatchMainserverEventsPoiRequestMock.mockReset();
+    dispatchMainserverEventsRequestMock.mockReset();
+    dispatchMainserverPoiRequestMock.mockReset();
     getWorkspaceContextMock.mockReset();
     withRequestContextMock.mockReset();
     createServerFunctionRequestDiagnosticsMock.mockReset();
@@ -82,7 +88,8 @@ describe('server transport', () => {
 
     const startFetch = vi.fn().mockResolvedValue(new Response('ok', { status: 200 }));
     dispatchMainserverNewsRequestMock.mockResolvedValue(new Response('news', { status: 200 }));
-    dispatchMainserverEventsPoiRequestMock.mockResolvedValue(null);
+    dispatchMainserverEventsRequestMock.mockResolvedValue(null);
+    dispatchMainserverPoiRequestMock.mockResolvedValue(null);
     dispatchAuthRouteRequestMock.mockResolvedValue(null);
     createStartHandlerMock.mockReturnValue(startFetch);
 
@@ -95,12 +102,13 @@ describe('server transport', () => {
     await expect(response.text()).resolves.toBe('news');
   });
 
-  it('bypasses mainserver events and poi requests before auth routing', async () => {
+  it('bypasses mainserver events requests before poi and auth routing', async () => {
     vi.stubEnv('NODE_ENV', 'production');
 
     const startFetch = vi.fn().mockResolvedValue(new Response('ok', { status: 200 }));
     dispatchMainserverNewsRequestMock.mockResolvedValue(null);
-    dispatchMainserverEventsPoiRequestMock.mockResolvedValue(new Response('events', { status: 200 }));
+    dispatchMainserverEventsRequestMock.mockResolvedValue(new Response('events', { status: 200 }));
+    dispatchMainserverPoiRequestMock.mockResolvedValue(null);
     dispatchAuthRouteRequestMock.mockResolvedValue(null);
     createStartHandlerMock.mockReturnValue(startFetch);
 
@@ -108,10 +116,32 @@ describe('server transport', () => {
     const response = await mod.default.fetch(new Request('http://localhost:3000/api/v1/mainserver/events'));
 
     expect(dispatchMainserverNewsRequestMock).toHaveBeenCalledTimes(1);
-    expect(dispatchMainserverEventsPoiRequestMock).toHaveBeenCalledTimes(1);
+    expect(dispatchMainserverEventsRequestMock).toHaveBeenCalledTimes(1);
+    expect(dispatchMainserverPoiRequestMock).not.toHaveBeenCalled();
     expect(dispatchAuthRouteRequestMock).not.toHaveBeenCalled();
     expect(startFetch).not.toHaveBeenCalled();
     await expect(response.text()).resolves.toBe('events');
+  });
+
+  it('bypasses mainserver poi requests before auth routing', async () => {
+    vi.stubEnv('NODE_ENV', 'production');
+
+    const startFetch = vi.fn().mockResolvedValue(new Response('ok', { status: 200 }));
+    dispatchMainserverNewsRequestMock.mockResolvedValue(null);
+    dispatchMainserverEventsRequestMock.mockResolvedValue(null);
+    dispatchMainserverPoiRequestMock.mockResolvedValue(new Response('poi', { status: 200 }));
+    dispatchAuthRouteRequestMock.mockResolvedValue(null);
+    createStartHandlerMock.mockReturnValue(startFetch);
+
+    const mod = await import('./server');
+    const response = await mod.default.fetch(new Request('http://localhost:3000/api/v1/mainserver/poi'));
+
+    expect(dispatchMainserverNewsRequestMock).toHaveBeenCalledTimes(1);
+    expect(dispatchMainserverEventsRequestMock).toHaveBeenCalledTimes(1);
+    expect(dispatchMainserverPoiRequestMock).toHaveBeenCalledTimes(1);
+    expect(dispatchAuthRouteRequestMock).not.toHaveBeenCalled();
+    expect(startFetch).not.toHaveBeenCalled();
+    await expect(response.text()).resolves.toBe('poi');
   });
 
   it('bypasses diagnostics for non server-function requests in development', async () => {
@@ -119,7 +149,8 @@ describe('server transport', () => {
 
     const startFetch = vi.fn().mockResolvedValue(new Response('plain', { status: 200 }));
     dispatchMainserverNewsRequestMock.mockResolvedValue(null);
-    dispatchMainserverEventsPoiRequestMock.mockResolvedValue(null);
+    dispatchMainserverEventsRequestMock.mockResolvedValue(null);
+    dispatchMainserverPoiRequestMock.mockResolvedValue(null);
     dispatchAuthRouteRequestMock.mockResolvedValue(null);
     createStartHandlerMock.mockReturnValue(startFetch);
     withRequestContextMock.mockImplementation(async (_input, callback) => callback());
@@ -150,7 +181,8 @@ describe('server transport', () => {
     );
     const logger = { info: vi.fn() };
     dispatchMainserverNewsRequestMock.mockResolvedValue(null);
-    dispatchMainserverEventsPoiRequestMock.mockResolvedValue(null);
+    dispatchMainserverEventsRequestMock.mockResolvedValue(null);
+    dispatchMainserverPoiRequestMock.mockResolvedValue(null);
     dispatchAuthRouteRequestMock.mockResolvedValue(null);
     createStartHandlerMock.mockReturnValue(startFetch);
     createSdkLoggerMock.mockReturnValue(logger);
@@ -191,7 +223,8 @@ describe('server transport', () => {
     const startFetch = vi.fn().mockResolvedValue(new Response('prod', { status: 204 }));
     const logger = { info: vi.fn() };
     dispatchMainserverNewsRequestMock.mockResolvedValue(null);
-    dispatchMainserverEventsPoiRequestMock.mockResolvedValue(null);
+    dispatchMainserverEventsRequestMock.mockResolvedValue(null);
+    dispatchMainserverPoiRequestMock.mockResolvedValue(null);
     dispatchAuthRouteRequestMock.mockResolvedValue(null);
     createStartHandlerMock.mockReturnValue(startFetch);
     createSdkLoggerMock.mockReturnValue(logger);
@@ -227,7 +260,8 @@ describe('server transport', () => {
     const startFetch = vi.fn().mockResolvedValue(new Response('plain', { status: 202 }));
     const logger = { info: vi.fn() };
     dispatchMainserverNewsRequestMock.mockResolvedValue(null);
-    dispatchMainserverEventsPoiRequestMock.mockResolvedValue(null);
+    dispatchMainserverEventsRequestMock.mockResolvedValue(null);
+    dispatchMainserverPoiRequestMock.mockResolvedValue(null);
     dispatchAuthRouteRequestMock.mockResolvedValue(null);
     createStartHandlerMock.mockReturnValue(startFetch);
     withRequestContextMock.mockImplementation(async (_input, callback) => callback());
