+++ packages/sva-mainserver/src/index.server.ts
@@ -1,3 +1,7 @@
 export * from './server/errors.js';
+export * from './server/events-route.js';
+export * from './server/interfaces-contract.js';
+export * from './server/news-route.js';
+export * from './server/poi-route.js';
 export * from './server/settings.js';
 export * from './server/service.js';
